+++ app/mailers/cancel_reservation_mailer.rb
@@ -1,0 +1,10 @@
+class CancelReservationMailer < ApplicationMailer
+  def cancel_reservation_email(reservation)
+    @reservation = reservation
+
+    mail(
+      to: @reservation.user.email,
+      subject: 'Your seance reservation has been cancelled'
+    )
+  end
+end
