+++ app/views/cancel_reservation_mailer/cancel_reservation_email.text.erb
