+++ app/views/confirm_reservation_mailer/confirm_reservation_email.text.erb
