--- spec/mailers/previews/reservation_mailer_preview.rb
@@ -1,10 +1,0 @@
-# Preview all emails at http://localhost:3000/rails/mailers/reservation_cancellation_mailer
-class ReservationMailerPreview < ActionMailer::Preview
-  def cancel_reservation_email
-    ReservationMailer.cancel_reservation_email(FactoryBot.build(:reservation))
-  end
-
-  def confirm_reservation_email
-    ReservationMailer.confirm_reservation_email(FactoryBot.build(:reservation))
-  end
-end
